requirements.txt
fastapi
uvicorn
pandas
numpy
matplotlib
seaborn
faiss-cpu
sentence-transformers
chromadb
transformers
